Factor diagnostic-specific XCTest support
